requirements.txt
streamlit
pydantic
requests
fastapi
uvicorn
chromadb
sentence-transformers
google-generativeai
python-multipart
beautifulsoup4
pymupdf
